Refactor licence enforcement
Remove mandatory license_docs attribute from install(), which also lets
us remove special-casing of "drake" targets (which shouldn't be in the
non-Drake-specific install.bzl). Implement a new check_licenses() rule
that consumes a list of install labels and verifies that each one
installs license files. Update various BUILD files accordingly.

This allows us to remove a lot of duplication from externals that have
multiple install rules, as now only the "master" rule for each external
needs to install the license files. This is also better as it allows us
to enforce checking for Drake itself and for VTK, which were previously
exempt.