Add mechanism to override distro detection
Add ability to override the built-in distribution detection for VTK and
Director. This makes it possible for users on unsupported platforms to
force use of one of the supported packages of these externals. (Needless
to say, this is an "if it breaks, you get to keep the pieces" deal.)

As a happy side-effect, this also refactors the shared distribution
detection code to be an external helper function.