Switch pydrake to use libdrake.so
Prior to this patch, each _pydrake_foo.so had its own copy of e.g.
drake/common/symbolic_variable.o object code.  That means that
global data was being duplicated / instantiated multiple times,
but Drake assumes that each global variable is a process-wide
singleton, e.g. for assigning each symbolic::Variable an ID.

This patch switches all of the _pydrake_foo.so to depend on the
libdrake.so rollup library, which will then be loaded just the
once into the Python interpreter.

In order do this, we need a header-only Bazel cc_library that
provides the headers for libdrake.so; we do that by resurrecting
some code removed in 31f79b63f79289de736218f067b7c1f5b27668d9
to transitively read out the headers based on the cc artifact.

Because the correct way to write a_pydrake_foo.so binary is now
very finicky, we factor their declarations into a helper macro.