Add unit test to catch port-level detection
There is one test that has a cycle, but no algebraic loop. That depends
on a system with no direct feedthrough to break the apparent loop.

I needed a test with direct feedthrough, but still doesn't have an
algebraic loop. But all of the systems *have* direct feedthrough. However,
the ports are wired in such a way that there is no feedthrough.
In this version, attempting to build the diagram throws
an exception. Once implemented, this test should fail.