Fix license check in case of "dummy" installs
Tweak license check to first verify if an install label actually does
anything, and to implicitly accept empty install labels. (An install
label that doesn't actually install anything is presumably okay not
installing a license, since there is nothing to be licensed.)

In particular, this affects VTK, as on macOS, we rely on VTK being
installed via homebrew, and the install label does not actually do
anything (it is merely present so that other code referring to it does
not need to be made conditional).

Also, add some comments around this logic.