Eliminate the "sorted" systems from Diagram
Because the systems are no longer, in any sense, sorted, the last vestige
of this sorting has been removed. As such, it is enough simply to reference
the owned systems. This also led to the collapse of Diagram::Own and
Diagram::Initialize into a single method (with ownership of systems passing
smoothly from blueprint to diagram.