Move protobuf support in drake/common into subdir
Bazel's support for protobuf does not have good backward compatibility.
As such, its helpful an important build file (drake/common/BUILD) didn't
mention @protobuf in a load() stanza.  Thus, we'll move the protobuf
support into a different package folder, so that only folks that truly
need it have to pay for it with their WORKSPACE entries.