Move PoseBundle::ToAutoDiffXd to testonly; add symbolic
This method was only used in the unit test as some convenience sugar.
It makes sense to transmogrify systems -- we need to change the memory
layout of the Context storage, the signature of port types, etc.

I'm not really sure it ever makes sense to transmogrify values, at least
not without AbstractValue providing a way that is generically usable.

But I am fairly sure that demanding that PoseBundle<symbolic::Expression>
be able to transmogrify into a PoseBundle<AutoDiffXd> is nothing we ever
want to support, so at a minimum the existing method signature was wrong.

If we decide we need this for some reason, we can move it back out of
the unit test code into some free function or other phrasing, once we
better understand what the purpose is.