Resolves #6461.
DiagramBuilder conservatively detected algebraic loops by examining systems and considering the possibility of an algebraic loop if the system had any direct feedthrough.

This changes the analysis to actually follow port dependencies. So, it's possible to have a system consisting of systems that all have direct feedtrhough, but because the ports aren't connected in a way that depends on that feedthrough, there are no algebraic loops.

Also, Diagram and DiagramBuilder had a relic of "sorting" the systems that has not been valid for a while. This also removes the (now impossible) sorting algorithm, expectation, and code artifacts.

The changes are as follows:

Add unit test that for the case where the old code would detect a loop, but the new code does not: diagram_builder_test.cc
Update loop algorithm:
diagram_builder.h - the actual algorithm
system.h, diagram.h, leaf_system.h (and their tests) - provide new API so that all pairs of input-output ports on a single system with a direct-feedthrough connection can be reported efficiently. Necessary to implement the port-level loop detection.
Removal or sorting artifacts. This includes modifying initialization of Diagram from DiagramBuilder. Affects: diagram.h and diagram_builder.h