Install examples and manupulation data
New `install_data` macro defined. This macro creates 3 filegroups:
* prod_models: data files in the current folder and its subfolders. Test files
(files contained in a `test` subfolder or named `*test*` are not installed).
* test_models: Only files contained in a subfolder named `test` or files named
`*test*`.
* models: prod_models + test_models

An install rule that copies these files in the corresponding install directory
is also created. The corresponding install directory is generated based on the
current folder which is prepended with "share/drake/":
The file `drake/drake/examples/acrobot/Acrobot_no_collision.urdf` will be
installed in:
{drake_install_dir}/share/drake/drake/examples/acrobot/Acrobot_no_collision.urdf