Work around possible Bazel providers bug
Apparently in newer versions of Bazel (starting with 0.5? As of 0.5.3 at
least...), Bazel doesn't recognize that a provider is the same entity if
one BUILD loads the .bzl file that declares the provider with a
workspace-qualified name, and another omits the workspace qualification.
Since these are the same .bzl, this feels suspiciously like a Bazel bug.
Anyway, work around the issue by always using the workspace-qualified
path.

This should get `bazel build install` working again with Bazel 0.5.3.