Change ToAutoDiffXd and ToSymbolic to throw when unsupported
Add -Maybe variant that returns nullptr when unsupported.

We used to DRAKE_ABORT_MSG but that got lost recently; this puts the
nice error back, this time as an exception to be nicer to users.
It also extends the nice exception to hold for symbolic conversion.

While we're here, substantially clean up the implementation.
The static method formulation no longer needs explicit choice of type.

Fix the static method formulation to set_name, like the instance
method already did.

Throw bad_cast if DoToFoo is broken; don't just return nullptr.