Reimplement hash functions used in maliput's GenerateObj
As a demonstration of the power and wisdom in the style-guide amendment
proposed in #6937, the lousy "something I found on the internet" composite
hash functions for GeoVertex and GeoNormal have been rewritten using the
machinery provided by `common/hash.h`.

The hashers have not been named `std::hash` here because they are only
intended to be used specifically for the indexing of vertices and normals
being assembled into an OBJ file.