Use the system protobuf.
This means using both the system version of the protoc compiler
and the system libprotobuf.so to link against.  To do this,
we copy a version of the protobuf.bzl file from upstream protobuf,
and modify it to use "command" instead of "executable" to run
protoc.  We then use pkg-config to find the libprotobuf.so
flags as appropriate.  The combination of all of this seems
to make things work properly.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>