Improve how to find drake resources
The function `FindResource()` and all the functions that rely on it search for
drake resources in the current directory and in a directory that is specified
by the environment variable `DRAKE_RESOURCE_ROOT`. However, in certain case,
the caller has an idea of where the resources are and should be able to
give additional paths to `FindResource()` where to seach. This is for
example the case when calling `FindResource()` from a python function from
pydrake such as `getDrakePath()` which is defined in a file installed in drake'
install directory. Resource paths can be added with the function
`GetResourceSearchPaths()` which adds the given path in a persistent variable.
This allows to get Drake's path and Drake's resources path from Python without
having to set the environment variable `DRAKE_RESOURCE_ROOT`.