Force using `python2` executable in testCommonInstall.py
On MacOS, using the executable named `python` might result in running the
system `python` which could lead to the following crash:

Fatal Python error: PyThreadState_Get: no current thread

Using `python2` instead of `python` forces the usage of the brew installation
of `python` against which `pydrake` has been built.

On Linux, `python` is typically a symbolic link to `python2` which is a
symbolic link to `python2.7`.