Switch to protobuf version 2. (#7304)
In order to switch to using the system version of protobuf
on Ubuntu 16.04, we need to switch back to protobuf 2 semantics.
The good news is that libprotobuf version 3 supports generating
code for protobuf version 2 files, so we can do this switch
now before we switch to the system version.  However, note that
the on-disk/on-wire protocols are *not* the same between version 2
and version 3, so any stored configurations will no longer work
after this change.