Avoid DRAKE_DEMAND failure on CheckStructuralEquality
This failure is improbable (but not impossible) for a user to hit, because
it requires a 64-bit hash collision in Formula::EqualTo, for two objects
that are both PSD formulas but have differing sizes.

(This was discovered while working on a hash_append spike test, that
removed the hash code short-circuit on equality checking.)

No unit test is added here, because we have to break some abstraction
barriers to tickle the error.  If we ever remove the hash short-circuit,
then the current unit tests already found the error.