Generate the cmake files for system libccd properly.
Since we now want to use the system version of libccd,
we need to generate an cmake file that properly finds it
(since the package on both Linux and macOS do not contain
one).  To do this, we add a new piece of functionality to
the pkg_config module, since pkg-config contains all of
the data we need to generate it.  Tested against drake-shambhala
on both Linux and macOS.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>