Fixup library paths for executables
Executables need to be fixed up to find where the libraries built
in the drake project are. Similarly to what is done for the libraries, on MacOS
the executable dependency paths are updated to correspond to the install tree
paths, and on Linux a relative RPATH is added to each executable to point to
the directory(ies) in which the libraries they depend on are.