Revert "Move symbolic headers back to hdrs"
This reverts commit 4d73b883ae4886340358320741bc791bc2f381c5.

As it turns out, what I thought was a rare pattern of putting headers
not into hdrs is actually fairly common within Drake, so I've made a
general solution to it in my 6996 feature branch.  With that in hand,
we don't need this work-around anymore, so we should go back to the
prior, superior phrasing.