Remove vtk from drake.cps.
Per the comments in https://github.com/RobotLocomotion/drake-shambhala/issues/38
VTK was listed in drake.cps to make sure that it is linked in the event
libdrake.so is not explicitly linking to it. But the config referenced below [1]
does indeed explicitly link VTK.

Removing this resolves errors with https://github.com/stonier/drake_pcl finding
Drake's VTK CMake config rather than the system CMake and not linking correctly.

[1]: https://github.com/RobotLocomotion/drake/blob/bdf86c07394b310a1f07b4e024fa3b84aed43d99/tools/install/libdrake/BUILD.bazel#L109