Resources can be found in install tree based on libdrake.so location
When an executable depends on `libdrake.so`, it can find the resources in the
install tree based on the location of `libdrake.so`. The implementation to
find `libdrake.so` is platform specific (MacOS and Linux). In both cases, it
finds the location of the library loaded in the process. It extracts its
directory and appends its with the relative path to the resource files.
This should limit the necessity of programmatically add path to the candidate
directory in which the sentinel file is searched.