Exclude the installation of vtk_util.h (#7258)
This PR is motivated by the need of supporting the case of a system providing VTK6 (Ubuntu Xenial default one) and Drake using the current VTK8 version. Since both VTK6 and VTK8 don't conflict in the binary symbols present in their public ABIs the risk present would be to mix both libraries in a compilation unit with vtk headers present in the mix. There is no guarantee about which version will be chosen since both libraries have the same names for headers.

The only header in Drake including vtk headers is vtk_util.h and this file is only used in .cc files inside Drake. This PR excludes the vtk_util.h file from being installed in the system so Drake users are safe about not including any vtk header at all.