Tidy up drake_cc.bzl
Tidy up defaults handling to prefer empty lists instead of None.
This tends to make macro bodies read more simply.

Remove hdrs forwarding from drake_cc_binary.  The hdrs= is not
allowed to be non-None on native.cc_binary, so we shouldn't be
fooling our callers that we support it.