Move the resource sentinel file down one folder
This changes the internal mechanisms of FindResource and its sugars
but does not affect users at all.  The actual data files still end
up in the same places; the relative_path args to FindResource are
unchanged; the GetDrakePath result is unchanged.

The shift is in order to enable 6996, which will move the entire
contents of drake-distro/drake up a level, thus moving the sentinel
back down to the repository root.