Collect our own list of transitive headers
Use our own list instead of the built-in `cc` provider so that we can
adjust and filter what is going to be installed, versus everything
that is required to compile.  This lets us move the current yaml and
vtk exclusions to their point of declaration.

This mechanism will also become important for 6996, where we have to
be very careful about include paths while compiling pydrake *.so's.