Fix -inl.h related header include order
When a component uses the -inl.h pattern, the related header include in
its .cc file is the -inl.h file.  Therefore, the related header include
in the -inl.h file must be the related .h file, so that we prove the .h
file parses on its own.

As it stands on master now, clang-format-includes allows both locations
without complaint, but after 6996 will force the headers down into the
group 5 mix; thus, we have to put suppressions in every -inl.h file to
allow this positioning.

If the -inl.h pattern accumulates, we can look into patching the tool
to allow (and require) this ordering without annotations.