Teach Bazel to conditionally handle packages hoisted up a directory
This prepares the Drake tree to move all of its source code up one
folder.  We add conditional wrappers to rewrite package paths from
//drake/foo:bar to //foo:bar, depending on whether Drake's skylark
global HAS_MOVED_6996 is true.

The script tools/dev/69960-move contains the code changes that will
be pushed in a subsequent PR.

Most of this PR should be reverted after the 6996 transition is
finished.