Installation and test issues of executables on MacOS
* fixup process for executables on MacOS was broken. `install_test` in
lcm allowed to expose that bug.
* `install_test` can take a little more than 300s to run in debug. This test
is changed from `medium` to `large` to avoid time outs.
* speed up install process by skipping all header files when checking if
file being installed is an executable.

For reference, to check the number of files with different extensions on
linux, one can run:
`find . -type f | sed -n 's/..*.//p' | sort | uniq -c`
in drake's install folder.

The result of this command is:
9 0
1 0/README
193 1
31 a
1 BSD
3 cc
195 cmake
1 cpp
2 csv
65 dae
1 doc
1 drake-find_resource-sentinel
1 glade
1 GPL
4283 h
60 hh
137 hpp
1 htm
22 hxx
11 in
18 jar
11 jpg
4 json
1 lcmtypes
1 ldpc
1 LGPL
1 MINPACK
1 MPL2
1 mtl
265 obj
29 png
12 proto
478 py
1 README
1 rst
29 sdf
332 so
56 stl
101 txt
1 TXT
23 txx
57 urdf
28 vtm
89 vtp
7 xml

Heuristically skipping header files is reasonable based on this result.