Make the protobuf fixup command work on macOS.
It had two problems.  The first problem was that the version
of bash on macOS is too old and doesn't understand ${var::-6}.
To get around that, we switch to ${var%??????}, which accomplishes
the same thing in a slightly uglier way.  The harder bit is
that macOS has a really ancient (or non-GNU; I'm not sure) version
of sed, and it really didn't understand the syntax that we
were using.  Instead, switch over to using awk instead, which
gets us the same results.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>