Speed up installation (#7683)
* Speed up installation

* Run `file` command only once, to detect if installed files are executables.
Instead of being run once for each file to test, it is run once with the list
of all the installed files, and returns the list of all installed files with
their file type. The list is then processed in line by line to assess if each
installed file is a binary executable or not. This reduces the installation
time by about 3 to 4s (out of 17-18s)
* Skip testing if a file is an executable if the extension belongs to the
following list: *.h, *.py, *.obj, *.cmake, *.1, *.hpp, *.txt. Since there
are more than 100 files installed for each of these extensions, it decreases
the installation time significantally (by about 3 to 4s).

Installation time is decreased from 16-20s to 8-10s.