Changes for newer ignition-rndf and ignition-math.
Here we update to a newer version of ignition-math4 and a
newer version of ignition-rndf (since they end up being
coupled).  Along with that, we fix up the ignition-rndf
build file to look substantially like the ignition-math
one for consistency.  While we are at it, we make
ignition-rndf a shared object.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>