Use yaml-cpp from system on Xenial and homebrew on MacOS. (#7602)
* Use yaml_cpp from Ubuntu Xenial.

This removes the bazel external yaml_cpp and replaces it with system
yaml_cpp found via pkg-config.

Using the system version cuts down on the size of libdrake.so and
improves linking compatibility when using Drake within systems that
use the system yaml-cpp version.

This change was previously introduced in #7352 and later reverted
(#7403) due to issues with the MacOS build. Other portions of that
original pull request were reintroduced in #7453.