Use stx and tinydir from github
Using new_local_repository with a workspace-relative path works alright
most of the time but has some sharp edges when used downstream.  For
things that are anyway on github, we might as well just download the
source archives rather than vendor the files.  (We had avoided this
earlier because adding them as git submodules to the CMake superbuild
didn't seem worthwhile at the time.)

This also upgrades both to the newest upstream, while we're at it.