Configure C and CXX compiler wrapper to compile project without using CROSSTOOL
The compiler wrappers are configured with CMake. This allows to compile drake
with CMake with clang 4 and 5. Without this wrapper, it is required to use
the CROSSTOOL file to manually specify `cxx_builtin_include_directory` since
there is a bug, either in clang or in bazel, that results in comparing
symlink directory names to canonical directory names and fails even though
the directories are technically the same (string comparison failing).