Remove the rename of liblcm-java.jar to lcm.jar.
The problem is that lcm.jar is not a proper rule, and hence the
rule in drake-lcm-spy that generates the launcher doesn't put
"lcm.jar" in the list of classes.  This ends up making drake-lcm-spy
fail to launch, since it can't find the main class.  Just remove
the rename, since it doesn't seem to be a problem to do that.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>