Remove the workspace entry for find_protobuf_cmake
Using a workspace to identify these files isn't terribly useful,
because there isn't great cause for allowing Drake downstreams to
replace them with a different copy.  Using new_local_repository
is a bit tricky to get right so we are trying to avoid it, thus
its worth just dropping this down to package-level dependencies
instead of a workspace.