Handle /path/to/*.so libs from pkg-config
Some *.pc files are odd (probably buggy) by declaring the --libs to be,
e.g., /usr/lib/x86_64-linux-gnu/libopencv_core.so.  We rewrite those to
be just, e.g., -lopencv_core so that they can make it to the cc_library
in at least some usable form.

We could try to add the literal *.so's to srcs instead (after adding
appropriate symlinking), but at least for opencv the -lfoo flags are
already (also) in the --libs output, so this is at least consistent now.