drake-lcm-spy install rule knows all renamed targets and launcher accepts input
When creating a java launcher, the install rule needs to know all the install
rules that it depends on that rename jars. Because of commit 6eca6e7 (libbot
removal), drake-lcm-spy was unaware of the fact that `liblcm-java.jar` and
`liblcmtypes_bot2-core-java.jar` were being renamed.
Additionally, the java launcher created at install time was not forwarding the
input arguments to the java application.