New bazel macro to install and test run executable from install folder.
`drake_install_and_test()` is a wrapper macro that creates an install rule
and creates a Python script that is run during the testing phase. The Python
script installs `drake` and runs the executables that are listed in the
install rule.
The executables can either be run normally, and the test expects them to
finish running and return 0 as their return code, or they can be started
and killed. The latter option allows to test executables that runs
continuously and do not provide the user a command line interface to
interact with them.
The Python script that is created inherites from all its dependencies
and therefore this macro should be used only once. This allows to
install `drake` only once, and by extension save time during tests.