Fix PopulateUpstreamToDrake for checkout not named drake
The prior version relied on "/drake/" appearing in absolute pathnames to
URDF files, then looked for package.xml files in the trip from the URDF
down to drake's root.  This worked when all resources were in
drake-distro/drake.  When we got rid of drake-distro, it still worked if
your git clone was named drake, or if you were running a test in the
bazel sandbox where the workspace is named drake.  If you were running
from the command line, it would give up because it didn't know where
drake was.

The fix is to use GetDrakePath to establish the coordinate to search up
until.  This still might fail if the URDF is from some foreign Drake
tree, but GetDrakePath is finding the current tree (it won't add the
foreign packages), but at least this is an improvement.