Allow and prefer "import drake.stuff" in python
This finally tips the scale on disambiguataing drake-the-workspace
from drake-the-lcmtypes in the python sys.path.  We have been fudging
it with careful deps= values so far, but it really is annoying to
debug.  By writing a custom __init__.py, we can fix it correctly
forever.

For now, we just repair tools.lint to use the newer import spelling;
in future PRs we can update the rest of the codebase.