Hoist up installed resource sentinel to match data
Now that the resources have been moved one directory up in drake's
install folder, we need to also move the resource sentinel.

Since we also have to change the installed location of resource_tool
(it had a double "drake/drake" in it), we take the opportunity to
relocate it under share/ to be alongside the other installed binaries.

As a smoke test of the move, add a regression test of kuka_simulation to
prove that it works once installed.  We'll plan a rephrase of the "test
of the install" logic soon, but for now this fixes a big problem, and
proves at works for at least one example program.

Co-authored-by: Francois Budin <francois.budin@kitware.com>