Initial version of the RigidBodyPlantBridge (#8027)
* Initial version of the RigidBodyPlantBridge

1. New class, allows a RigidBodyPlant to contribute to GeometrySystem
   (currently to a limited degree).
2. Modify examples/contact_model/bowling_ball.cc to illustrate the usage.

* Move rigid body plant bridge into multibody/rigid_body_plant

Move the files, change the references, change namespace.

* Address reviewer comments

* Debug build failure

RigidBodyPlantBridge was declared auto-diffable. However, the underlying
tree isn't. In debug builds (even though I'm not using an AutoDiff version)
it is being exposed.

* Missed re-ordering of includes with movement of bridge

* Fix Calc**Ouput methods

Based on the newly understood semantics that I can't take for granted
what the state of the input parameter is.