Disable pardiso support in OSQP
OSQP's support for pardiso involves dlopen'ing a pardiso shared library;
in general, Drake does not have access to such a thing, so this code had
no practical effect.  However, it did have the downside of adding a
dependency on dlopen, and worse not declaring `-ldl` as a build-time
dependency.  Instead of adding `-ldl`, we choose to disable pardiso
until such a time as we need it, and have CI support for pardiso.