Add ability to alter the mirrors used in workspace
The tools/workspace/default.bzl gains a `mirrors` argument that allows
changing which mirrors are used.  The argument is passed down to only
the repositories that use mirrors; to ensure that's reliably achieved,
all of the individual rules and repositories declare mirrors = None as
an argument and fail if no value was supplied.  Only the default.bzl
provides a default value.