Fix Wshadow warnings from GCC (systems)
This removes potential confusion between member fields and local
variables.  We are careful to leave the Doxygen spellings alone
(not infected with -in suffixes).

For DynamicProgrammingOptions, we bury a its constructor into the cc
file, since its constructor is unlikely to be performance-critical.

For TimeVaryingData stuff, we out-of-line the constructor definitions,
which lets us use different names for declaration vs definition, but
does not affect compiler inlining operations.