Fix Wdeprecated warnings (geometry)
When using clang, Wdeprecated warns about deprecated implicitly-defined
copy constructors (happens most often due to user-defined destructor):
http://en.cppreference.com/w/cpp/language/copy_constructor#Implicitly-defined_copy_constructor

In ShapeToLcm we were doing DRAKE_DEFAULT_COPY_... for some reason,
which thus generated a copy-assignment operator (that we never used),
which thus called the default ShapeReifer copy-assignment operator,
which is deprecated because of the ShapeReifer (virtual) destructor.

There is no reason to ever copy a ShapeToLcm, so we'll mark it
DRAKE_NO_COPY... and thus not generate any deprecated operators.