Fix Wdeprecated warnings (multibody/shapes prep)
In a future patch where the copy constructors on DrakeShapes classes are
demoted to private, we see that these lines would no longer compile
because they are (via auto) taking a copy of the DrakeShapes classes,
instead of merely capturing the downcast reference.

This patch fixes all such assignments to capture the downcast reference,
as was likely intended by the original author.