run_and_kill() was always prepending command line
`run_and_kill()` is a helper function that runs a command and kills it after
a set time. In the test implemented to test this function, we run a command
that is on the system, but that is not part of the installation tree.

This commit also forces the command to be run from the current working
directory '/' to avoid false positives (e.g. resource files found in
the source tree instead of the install tree).