Fix Wdeprecated warnings (common/trajectories)
When using clang, Wdeprecated warns about deprecated implicitly-defined
copy constructors (happens most often due to user-defined destructor):
http://en.cppreference.com/w/cpp/language/copy_constructor#Implicitly-defined_copy_constructor

This patch explicitly defines the copy, move, and assignment semantics
and visibility of the virtual class hierarchy used by the trajectory
classes.  Thus, the warning is silenced and the code is safer.

This also changes all Clone() implementations to use the copy
constructor (and be defined in the .cc file for consistency),
instead of listing out member fields by hand.

We clean up some related items while we're here:
- Move private members to end of class.
- Order ctor/dtor/clone class declaration per cppguide.
- Use `= default` for all defaulted ctors and dtors consistently.
- Use drake::MatrixX spelling consistently.
- Use west-const and ref-with-typename more consistently.
- Repair a few doxygen nits.

We stop using eigen_aligned_std_vector, because it serves no purpose on
our supported platforms.  Best to get this API change in as this class
is already being refactored, before other code starts to use it.