PR Fix Wdeprecated warnings (common/proto)
When using clang, Wdeprecated warns about deprecated implicitly-defined
copy constructors (happens most often due to user-defined destructor,
but in this case is due to the explicit assignment operator):
http://en.cppreference.com/w/cpp/language/copy_constructor#Implicitly-defined_copy_constructor

This patch explicitly defines the desired copy constructor, silencing
the warning in call_python.

We also document the public (but internal) methods in order to reason
about why exposing the copy constructor is correct.