Add LCM Publish method that both encodes and sends (#8425)
PR 7178 had many unfixed defects.
This is a first downpayment on fixing them.

A virtual method cannot have a default argument per GSG, but that PR added
one on Publish. Instead of fixing all of the call sites to pass 0.0 explicitly, we port
a major portion of them to some new sugar that both serializes and publishes,
which is a nicer approach anyway.

This also sets the stage for 8256 where subscription-related sugar methods will
be introduced.