Add unzip to source distribution install_prereqs script
Technically only CI needs this, but since we require zip, I am going to argue it is not too controversial to require unzip, and saves me having to document the requirement elsewhere. In addition, SNOPT seems to be most widely distributed as a zip-file, so this may well be needed in future anyway.