Merge id port and pose port into single port
Previously, the connection between geometry source and GeometrySystem was
a *pair* of ports: id and pose. The idea was that ids would be reported
once and each kinematics value (pose, velocity, acceleration) would come in
on a coordinated port. This leads to coordination challenges and two ports
in even the simplest case.

This is the beginning of the trend where each kinematics data type will
come with its own frame ids. This leads to redundancy but simpler looking
code.

This provides a very simple, generic API for writing pose data to the
output port's abstract value: clear and set each frame. Subsequent PRs
will add sugar to facilitate this. But this initial interface is sufficient
in that everything that needs to be done *can* be done.

Changes are:
1. Removal for frame_id_vector.* (and tests)
2. Modification of FrameKinematicsVector to include FrameIds.
3. Update of GeometrySystem and GeometryState to consume the new
   FrameKinematicsVector in place of the pair of inputs.
4. Modifications of all existing examples that connect to GeometrySystem.