Fixes inconsistency with teh default values of basic vectors and named (basic) vectors.
Default values of named vectors should use dummy_value<T>... they were simply being set to zero.
This was inconsistent with the behavoir of every other BasicVector.

This was particularly nasty for symbolic::Expressions, which were simply being set to zero by default
(and therefore getting simplified right out of the equation).

Also added a helper to let named vectors get initialized to a Variable with the known variable name.  (this makes using named vectors with Expression so
much nicer).