Restore behavior of Box shapes in BulletModel.
Updating Bullet to 2.87 introduced a fix that caused sub-epsilon
differences in the vertices returned by `btBoxShape::getVertex()`.
SNOPT exhibits different behavior when run with constraints that use the
slightly different set of vertices introduced by the fix.

We were only using `btBoxShape` when adding boxes to the collision model
(the collision shape that was stored in the model was always a
`btConvexHullShape` due to an (unrelated?) issue with collision normals
in `btBoxShape`). It was basically a lazy way to convert box dimensions
to vertex coordinates. This PR explicitly converts the box dimensions to
vertex coordinates and locks down the behavior we expect when creating a
`btCollisionShape` from a `DrakeShapes::Box`.