Add "multiple choice formulation" to approximate a bilinear product with its McCormick Envelope (#8655)
Add an alternative formulation to relax w = x * y through mixed-integer convex formulation. The feasible region of this alternative formulation is identical to the SOS2 formulation, i.e., the convex hull of the surface w = x * y in each region. But according to Vielma's 2010 paper, this formulation might be faster for mixed integer problem with small number of segments.