Update lcm-related bazel library names
Remove //util:lcm_util forwarding header, long since deprecated.

Rename //lcm:lcm to //lcm:real to clarify that it is only a concrete
implementation of the interface, instead of everything in the package.
Do not rename in-tree uses to match; they can use the package-level
library without adding overly many spurious dependencies.

Rename //systems/lcm:lcm to //systems/lcm:lcm_pubsub to clarify that it
is only the publisher and subscriber systems, instead of everything in
the package.  Rename all in-tree uses to match.

Add drake_cc_package_library for //lcm and //systems/lcm.