Better deal with -pthread in pkg_config_repository
Through some chain of pkg-config machinations that I cannot understand,
the gtk+-2.0.pc (on Ubuntu 16.04) ask for -pthread in --cflags but not
in --libs, even though all of its dependencies appear to list it in
either both places or none.  In any case, rather than insisting on
pkg-config being consistent, we choose to just use linkopts for
pthreads, always.