Rework OutputPort to not depend on System
This breaks the cycle between system.{h,cc} and output_port.{h,cc}, in
prepartion for moving the method bodies into the header in support of
user-defined scalar types.

Relatedly, we stop forward-declaring a Context<T> and instead directly
include it, per GSG rules.

Relatedly, we also stop casting nullptr to a reference to return; this
is general a bad idea, and will cause problems with the user-defined
scalar commit.