Remove stale comment
This comment dates from the very first prototype of the installation and
packaging, where the top-level install list had to explicitly call out the
license files for each external.  However, with the install.bzl rules that we
use now, where license data files are installed by the external itself, and
declared and enforced in the build system, the comment here no longer makes any
sense.  Indeed, the target it refers to no longer exists.