Introduces MBP public API for adding collision filters via bodies
Ultimately, this is the mechanism that enables the user to reproduce
an effect akin to "collision filter groups". Example (drawn from the
code):

std::vector<const RigidBody<T>*> bodies{&body1, &body2, &body3};
geometry::GeometrySet set = plant.CollectRegisteredGeometries(bodies);
scene_graph.ExcludeCollisionsWithin(set);

The user specifies a group of *bodies* and then asks the plant for a set
of geometry registered to that collection of bodies. Then that geometry set
can be passed into the SceneGraph API for filtering collisions. There is
no syntactic sugar but provides the foundational functionality.