Add API for users to create "collision filter group"-like filtering via MBP (#8954)
* Introduces MBP public API for adding collision filters via bodies

Ultimately, this is the mechanism that enables the user to reproduce
an effect akin to "collision filter groups". Example (drawn from the
code):

std::vector<const RigidBody<T>*> bodies{&body1, &body2, &body3};
geometry::GeometrySet set = plant.CollectRegisteredGeometries(bodies);
scene_graph.ExcludeCollisionsWithin(set);

The user specifies a group of *bodies* and then asks the plant for a set
of geometry registered to that collection of bodies. Then that geometry set
can be passed into the SceneGraph API for filtering collisions. There is
no syntactic sugar but provides the foundational functionality.

* Add short-term method for eliminating visual geometry from collision

All geometry registered *through* the MultibodyPlant as "visual" geometry
are now filtered from collisions. This is *not* the real solution (these
geometries are still being included in the proximity engine), but it
improves performance and eliminates erroneous contact results in the
interim.