symbolic: Allow for EIGEN_INITIALIZE_MATRICES_BY_ZERO
When a downstream user of Drake sets EIGEN_INITIALIZE_MATRICES_BY_ZERO,
Eigen will assume that any Eigen Scalar can be constructed from "0"
(a literal zero int).  This fixes the symbolic variable, monomial,
polynomial, and formula scalars to obey that contract.

Since an Eigen::Matrix over these types would be using the default
constructor when the ..._BY_ZERO macro is unset (as in Drake's default
build configuration), we choose to overload them to promote the zero int
to a nullptr, and then delegate to the existing default constructor.