Introduce penetrating pair order invariant guarantee (#9070)
If two geometries (g and h) are penetrating, SceneGraph (via its supporting
infrastructure) did not guarantee whether the pair would be reporting as
(g, h) or (h, g). While the values were correct, regardless of order, it
led to incoherency in interpreting them (particularly visually). This
introduces a guarantee that the pair will *always* be reported in the same
order (although the basis for the order is left undocumented.)