Reimplement Monomial::operator== without using std::map::operator==
Consider a case where we have `m1 = {x -> 1}` and `m2 = {y -> 2}` where both of
`m1` and `m2` are of `std::map<Variable, int>` which is the internal
representation of the symbolic::Monomial class. When `m1 == m2` is executed, it
builds a symbolic formula `x == y` and then convert it to `bool`. This triggers
`(x == y).Evaluate()` which will throw a std::runtime_error if
`Formula::Evaluate` does not use structural equality in handling
equalities. Note that this change will come in the next commit.

In this PR, we reimplement Monomial::operator== to avoid this problem. We are
explicitly using `Variable::equal_to`. As a result, no symbolic formula will be
created and evaluated. This change should boost performance as well.

This PR reverts https://github.com/RobotLocomotion/drake/pull/6596.